# Avoid use ArrayBuffer to create COO format matrix. 